+++ src/types/ErrorMessage.ts
@@ -1,0 +1,8 @@
+export enum ErrorMessage {
+  LoadTodos = 'Unable to load todos',
+  AddTodo = 'Unable to add a todo',
+  DeleteTodo = 'Unable to delete a todo',
+  UpdateTodo = 'Unable to update a todo',
+  UpdateTodos = 'Unable to update todos',
+  EmptyTitle = 'Title should not be empty',
+}
